dnsproxy: Filter DNS records based on IP families enabled
[dnsproxy] Filter DNS records based on IP families enabled. JB#48769

If a VPN is connected over a transport that supports both IPv4 and IPv6
and the VPN is IPv4 only, also AAAA DNS queries are sent, thus responses
may contain AAAA records if the DNS server supports them. In such case,
connections can be made over transport's IPv6 network bypassing VPN
connection.

The issue may get even worse if the SingleConnectedTech is not used and
there are multiple networks online at the same time. If network A and B
are both online and A supports IPv4 and IPv6, whereas B being the
default service supports only IPv4 and is used transport for a VPN, then
traffic is routed to A's IPv6 network if there is a AAAA record for
requested domain name.

Therefore, it is imperative to filter out the DNS requests and replies
that have no network available on the default service. This change makes
this behavior configurable, as in some setups multiple network types
and/or services are to be used in conjunction.

When the request type is filtered out the reponse is sent as
ns_r_noerror message containing the authoritative bit. No query is done
in this case.

If the DNS record is other than A or AAAA (e.g., CNAME), it will be
processed. To be able to detect the type of the DNS request/reply
cache_check() is changed to always set the type before returning in case
the type is something else than A or AAAA. This allows to filter out the
A/AAAA records when the necessary IP family is not connected on the
default service, and also allows to return a reply when the type is not
A/AAAA, e.g., CNAME.

Default service and ipconfig notifications are utilized to set correct
IP family types enabled on dnsproxy. When the IP family is disabled it
is also removed from resolvfile, and when only either of the IPv4 and
IPv6 is used the additional options are enabled as well to the
resolvfile. This is to minimize the delays caused by applications
sending both A and AAAA requests as the currently used 'single-request'
forces the requests to be sequential.

Also, replace all static 1/5/28 use for DNS record types with correct
types from arpa/nameser.h.