wifi: Detect invalid key with 4-way handshake offloading
wpa_supplicant is able to offload the 4-way handshake into the driver.
In this case the 4WAY_HANDSHAKE state is skipped and therefore the
current mechanism to detect a wrong password doesn't work any more. This
patch extends the detection mechanism to also work with the offloading
feature. The status code 16 is relatively generic, so this can lead to
false positives.