bluetooth: Move DBG statement after initial checks
gcc complains with format-overflow for the DBG statement for the
bridge argument. Let's move the DBG after the initial checks to avoid
the complain and also make the output more helpful. If we see it in
the log, the bridge has been created.

(cherry picked from commit 8d0147d7578307ff25b0823cfdfc366c7e9eaafb)