resolver: Support options in resolvfile and prepending entries
[resolver] Support options in resolvfile and prepending entries. JB#48769

Add support for enabling options in resolvfile with a simple toggle
function. This currently enables 'single-request' option, which is
intended to be used when only either of IP families is in use. Reason is
that with this option the requests are sequential instead of parallel
and will eliminate the default delay. However, the queries take twice
the time to complete when this option is set so it is not to be set when
both IP families are supported. Many glibc applications do both A and
AAAA queries in parallel.

Added also support for prepending a nameserver entry to the top of the
list. This allows to put IPv6 nameserver on top of IPv4.