iwd: Filter out connect failure for auto connect mode
If the user ask explicitly to connect to a network and the operations
fails for any reason the ConnMan core will go from either idle or
connected to failure state. In failure state the core will call
connman_network_disconnect() which will be propagated to iwd.

ConnMan will set the AutoConnect flag on KnownNetwork on
bootup/discovery. iwd tries interpretes this as request to connect
now. If this is not possible, e.g. out of reach, iwd response with an
failure. This disables the auto connect mode of iwd.

So the simplest thing here is not to tell the core the connect fails,
instead just don't do anything if auto connect is enabled. This keeps
the auto connect of iwd working and ConnMan's state machine doesn't
care either.

(cherry picked from commit 499bfb619f2077ed9ba2a4aec18bcb5b745bc888)