wifi: Duplicate GSupplicantSSID pointer members
In ssid_init(), the pointer members of GSupplicantSSID *ssid was
initialised to point to members of a connman_network, which get
freed and replaced, causing ssid's pointers to point to freed heap
memory. To mitigate this, these members are duplicated and then
freed and along with ssid when no longer needed.

(cherry picked from commit 481d08f108fed108a6174ad00ef73393edd99139)