wifi: Handle invalid-key case on WPA-SAE authentication failure
On WPA3-SAE authentication, wpa_supplicant goes directly from
authenticating to disconnected state if the key was invalid.

The above is currently not handled and the `connect-failed` error is
reported on such cases. In order to make the client agent prompt for a
new password, we need to handle this transition and report the
`invalid-key` error.

(cherry picked from commit 0d5d05f2a6a6895aa2b1ffeda34489552468327b)