[connman] Start service after quota to avoid fs read-only issues. Fixes JB#61621
Set the drop-in which requires home to be mounted to order connmand (and
connman-vpnd, which depends on connmand) to run after the quota service
for home has been run. Quota service is triggered by home.mount and may
make the fs as read-only for the period of quota check. This is an
insurance to ensure that there will be no filesystem read-only issues
regardless of what the quota service does.