[connman] Cancel set_online request if modem is powered off. JB#61333
Had Android device paired via bluetooth to Sailfish one, then
disconnected them. A hfp modem was created which stayed offline.

That's a problem of its own, but amplified by connman trying every 2
seconds to make the modem go online. That's doomed to fail as long
as it's not powered. Thus, if set_powered fails, the online attempts
are stopped.