dnsproxy: zero-terminate the buffer after recvfrom
in the case of udp_listener_event(), recvfrom reads into a buffer
which is then passed down into:
  resolv()
  ns_resolv()
  ns_try_to_resolve_from_cache()
  cache_check()

which finally invokes strlen on the buffer.  It isn't clear whether the
read-in length is ever taken into account, so zero-terminate the buffer
for now.

Similarly, in tcp_client_event, the buffer filled by recvfrom is sent to
  read_tcp_data()
  cache_check()

(cherry picked from commit 16044e9fcad813f0a3576540448137803a19bffd)