tethering: Add more client details, track DHCP lease, add getter
[tethering] Add more client details, track DHCP lease, add getter. JB#60565

Add a separate "added" table for new MAC clients instead of signaling
all registered clients.

Add the type of the tethering client to be added with the register.
Values are service types. In addition to this track the DCHP lease and
signal it when changed, use the ipconfig method type as type. When
changed add the MAC to the "added" list and notify it as a D-Bus signal.

The D-Bus method to get GetTetheringClientDetails() returns a dict,
which has MAC address as key (string):
 - "Address"		IP address	(string)
 - "AddressType"	IP address type (byte - ipconfig method type)
 - "Technology"		Tech type	(byte - service type)
 - "Version"		Tech vesion	(byte - WiFi: 2/5, BT: LMP)

Add a getter for the list of clients registered to be used when checking
which clients left and which are new if backend does not support this on
WiFi/Bluetooth.

Drop the "__" prefix from the tethering functions. These are for
internal functions but the main use of these is by plugins.