wireguard: Rework and fix hostname resolving
[wireguard] Rework and fix hostname resolving. Fixes JB#63273

This change changes how the resolving of the defined host is done.

First, if the host is defined as an IP address there is no need to start
the DNS reresolve. This is checked at the connect steps.

Second, the endpoint parsing is split into two, mainly because the
resolving of the host is done differently when doing initial connect,
and when executing the DNS reresolve. This is because initial resolving
can be done with getaddrinfo() but reresolve is done with GResolv to
avoid blocking in case the network is not properly setup, which leads
vpnd to be frozen until the timeout occurs. With GResolv the results are
simply converted to sockaddr struct(s) using getifaddrs() with
AI_NUMERICHOST to avoid another DNS query.

Third, the local DNS servers are added for GResolv to use. I.e.,
localhost in both IPv4 and IPv6 (in the same order as in resolv.conf).

Fourth, the stop condition for DNS reresolve does now retry also with
name error (NXDOMAIN) and no answer (NOERROR without a message).