wireguard: Rework hostname resolve, split code and do not resolve IP
[wireguard] Rework hostname resolve, split code and do not resolve IP. JB#63273

This changes how the resolving of the hostname is done and restructures
the code.

First, the endpoint parsing is split into two, mainly because the
resolving of the host is done differently when doing initial connect,
and when executing the DNS reresolve. This is because initial resolving
can be done with getaddrinfo() but reresolve is done with GResolv to
avoid blocking in case the network is not properly setup, which leads
vpnd to be frozen until the timeout occurs. With GResolv the results are
simply converted to sockaddr struct(s) using getifaddrs() with
AI_NUMERICHOST to avoid another DNS query.

Second, if the host is defined as an IP address there is no need to
start the DNS reresolve.

Third, the stop condition for DNS reresolve does now retry also with
name error (NXDOMAIN) and no answer (NOERROR without a message).