service: Send the DNS servers of VPN's transport when VPN is ready
[service] Send the DNS servers of VPN's transport when VPN is ready. JB#63273

Add a new property to be sent from the VPN service's path when the VPN
reports being ready. This is sent similar to other DNS service lists
when they change but with name "TransportNameservers".

The reason for doing this is that WireGuard does the periodic DNS
reresolve after getting connected, and the reresolving needs to be
done using the DNS server(s) of the transport service, and not the ones
WireGuard has set up.