[connman] Use systemd-resolved, clean use of resolv.conf. JB#63311
Replace internal dnsproxy.c with systemd-resolved. This requires that
ConnMan does not set up the resolv.conf symlink to its internal
resolv.conf, or to any other location as it seems to prevent resolved
from restarting. To faciliate this fully, remove the resolv.conf symlink
if it points to the ConnMan resolv.conf at post action.