systemd-resolved: Use DNS servers of the default service
[systemd-resolved] Use DNS servers of the default service. Fixes JB#63341

Register to get the notifies about the default service, service state
and offline mode changes. When default service changes set it as
default route in resolved, revert other links' status and flush the cache
so only the default service's DNSs are used. This brings the similar
behavior that is is in dnsproxy.c, and prevents leaking of DNS queries
outside VPNs as well. Thus, effectively making only the DNS servers of
the default online service to be used.

To ensure that the servers are gone the domains and servers lists are
cleared when the service is disconnecting. The current default index is
being tracked and when it is being reset, the cache will be flushed as
well.