dnsproxy: Fix NULL/empty lookup causing potential crash
[dnsproxy] Fix NULL/empty lookup causing potential crash. Fixes JB#63425

In ConnMan through 1.44, the lookup string in ns_resolv in dnsproxy.c
can be NULL or an empty string when the TC (Truncated) bit is set in
a DNS response. This allows attackers to cause a denial of service
(application crash) or possibly execute arbitrary code, because those
lookup values lead to incorrect length calculations and incorrect
memcpy operations.

This patch includes a check to make sure loookup value is valid before
using it. This helps avoid unexpected value when the input is empty or
incorrect.

Fixes: CVE-2025-32743