[connman] Summarize changes for 5.0.0. JB#63495
The changes taken from ConnMan master to 5.0.0 include:
 a) conf.d support for changing existing configiration with
 b) systemd-resolved utilization as the DNS backend
 c) Option for making all new WLAN/Ethernet services to have mDNS on
 d) Synchronization of upstream dnsproxy.c including refactoring & fixes
 e) Fix for tethering client's DNS with systemd-resolved

More clarification on the features

a) ConnMan has now a conf.d support that allows to change the
configuration variables with added files to /etc/connman/main.conf.d for
connmand and /etc/connman/connman-vpn.conf.d/ for the connman-vpnd. The
files must have .conf suffix, and they follow the same syntax as the
existing configurations, all configuration options should be in
General section. The values set in additional files override the
existing ones.

b) systemd-resolved is set as the default DNS backend to serve the DNS
requests, and to provide caching and mDNS support. Other features may
be enabled later. This should work similar to the internal dnsproxy, to
use only the DNS servers of the default service used.

c) There is an configuration option to enforce mDNS for all newly
created services (WLAN/Ethernet), which does not affect existing
services. With DefaultmDNSConfiguration = true this can be enabled in
the configuration. Similar option exists for tethering, for the hotspot
the mDNS can be enabled with TetheringmDNSConfiguration = true.

d) This takes the latest dnsproxy.c code from upstream including all the
refactorings made and also the fixes done to improve security and
address CVEs.

e) This makes the DNS requests sent by tethering clients to the hotspot
using systemd-resolved as a backend to be proxied to the backend. This
is identical behavior to previously used internal dnsproxy and allows
systemd-resolved to process the requests, and to have the requests
cached by the resolved itself, also for the use of tethering clients.