[sailfish_iptables] Use prefix excludes when saving/restoring. JB#63544
Use the prefixes defined with IptablesPrefixesToExclude in configuration
to exclude the targets, chains and rules that have them. Those are to be
added by the component to which they belong and ConnMan should not take
part on restoring them, since the targets may not have been setup yet,
causing errors and maybe propagates to other bigger failures.