vpn-provider: Change vpn_provider_password_cb_t to use int error
[vpn-provider] Use int error in vpn_provider_password_cb_t. JB#47886

The use of this callback is in VPN plugins only, it is meant to be used
after the reply from agent is received. To reduce code and not to
require parsing the char string error message it is preferred to use the
vpn_agent_check_and_process_reply_error() to get the error code for the
message and use it for the callbacks.