pptp: Do not clear credentials on connection error
[pptp] Do not clear credentials on connection error. Fixes JB#63774

Utilize the authentication and connection errors to detect which error
should be reported back on the pppd error code 16 (modem hangup).
Connection errors will be reset when connection succeeds, or the service
is saved. If there is at least one authentication error code 16 means
the credentials need to be cleared.