l2tp: Implement shared secret support using file or VPN agent
[l2tp] Implement shared secret support using file or VPN agent. Fixes JB#63779

Use the "Challenge" option properly by implementing the shared secret
support either by 1) user defined file 2) system l2tp-secrets file or 3)
using VPN agent to ask the shared secret, local and remote hostnames
from the user. With 1) and 2) the file is used as is. With 3) the
secrets file connman-xl2tpd-secrets.conf is created in the VPN_STATEDIR
and is removed like the rest of the configuration files.

Use the "AuthFile" to specify a secrets file to use, or if it is defined
as "-" use the system file instead. If "Challenge" is enabled the
defined auth file is automatically used, if there is no auth file
defined VPN agent is used to query the user.

A new callback for setting the shared secret info is added that is always
executed before the credentials callback.  All information is requested
in the VPN agent query (5 strings total) if "Challenge" is enabled and
no auth file is defined.